requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pillow
google-generativeai
jinja2